Rename provider to repositorymanagementsystem
